remove history
