+++ base/telegram-trading-bot/src/utils/formatters.ts
@@ -60,23 +60,9 @@ export function formatAddress(address: string): string {
  */
 export function formatBalanceMessage(
   ethBalance: string,
-  tokens: TokenInfo[]
 ): string {
   let message = `💰 *Your Balances*\n\n`;
   message += `*ETH*: ${formatEthBalance(ethBalance)} ETH\n\n`;
-
-  if (tokens.length > 0) {
-    message += `*Tokens:*\n`;
-    tokens.forEach((token) => {
-      const formattedBalance = formatTokenBalance(
-        token.balance,
-        token.decimals
-      );
-      message += `${token.symbol}: ${formattedBalance}\n`;
-    });
-  } else {
-    message += `No other token balances found.`;
-  }
 
   return message;
 }
